Middle of a large refactor to use a single tenant schema database in Clickhouse instead of multiple databases.

+++ unify/file_adapter.py
@@ -16,20 +16,32 @@ class LocalFileTableSpec(TableDef):
         super().__init__(table)
         self.file_uri = opts['file_uri']
         self.reader_name = opts['reader_name']
+        self.options = opts.get('options', [])
     
     def query_resource(self, tableLoader, logger: UnifyLogger):
         path = Path(self.file_uri)
 
         size_return = []
+        kwargs = {}
+
+        for idx, k in enumerate(self.options):
+            if k == 'skip':
+                count = int(self.options[idx+1])
+                kwargs["skiprows"] = count
+            elif k == 'header':
+                count = int(self.options[idx+1])
+                kwargs["header"] = count
+
         method = getattr(pd, self.reader_name)
         if 'chunksize' in inspect.signature(method).parameters:
-            with method(path, chunksize=5000) as reader:
+            kwargs["chunksize"] = 5000
+            with method(path, **kwargs) as reader:
                 for chunk in reader:
                     chunk.dropna(axis='rows', how='all', inplace=True)
                     chunk.dropna(axis='columns', how='all', inplace=True)
                     yield AdapterQueryResult(json=chunk, size_return=size_return)
         else:
-            df = method(path)
+            df = method(path, **kwargs)
             yield AdapterQueryResult(json=df, size_return=size_return)
 
 
@@ -49,45 +61,68 @@ class LocalFileAdapter(Adapter):
             ]
         return self.tables
 
-    def list_files(self, path: str) -> list[str]:
-        if path is None:
-            path = ""
-        p = Path(os.path.join(self.root_path, path))
-        return [f.name for f in p.glob("*")]
+    def list_files(self, match: str) -> list[str]:
+        if match is None:
+            match = "*"
+        else:
+            match = match.replace('%', '*')
+        return [f.name for f in self.root_path.glob(match)]
+
+    def _resolve_path(self, path: str):
+        userp = Path(path)
+        if path.startswith("/"):
+            return userp
+        else:
+            return self.root_path.joinpath(path)
 
     def can_import_file(self, path):
         # Path will be whatever the user entered. Either it will be a path relative
         # to the system root, or it could be absolute in which case it needs to
         # start with our same root
+        userp = self._resolve_path(path)
+        return userp.is_relative_to(self.root_path) and userp.exists()
 
-        userp = Path(path)
-        if path.startswith("/"):
-            return userp.is_relative_to(self.root_path) and userp.exists()
-        else:
-            return self.root_path.joinpath(path).exists()
+    def peek_file(self, file_uri: str, line_count: int, logger: OutputLogger):
+        file_path = self._resolve_path(file_uri)
+        reader = self.determine_pandas_reader(file_path)
 
-    def import_file(self, file_uri: str, options: dict={}):
+        if reader == pd.read_csv:
+            with open(self._resolve_path(file_uri)) as f:
+                for x in range(line_count):
+                    line: str = f.readline()
+                    if not line:
+                        break
+                    logger.print(str(x+1) + " " + line.strip())
+            return None
+        elif reader == pd.read_excel:
+            df = pd.read_excel(file_path)
+            df = df.head(n=line_count)
+            df.insert(0, 'row', df.index + 1)
+            return df
+
+    def import_file(self, file_uri: str, options: list=[]):
         print("importing the file")
 
-        if not file_uri.startswith("/"):
-            file_uri = str(self.root_path.joinpath(file_uri))
+        file_path = self._resolve_path(file_uri)
+        if not file_path.exists():
+            raise RuntimeError(f"Cannot find file '{file_uri}'")
 
-        reader = self.determine_pandas_reader(file_uri)
+        reader = self.determine_pandas_reader(file_path)
         if reader is None:
             raise RuntimeError(f"Cannot determine type of contents for '{file_uri}'")
 
         # Now create an empty table record which we will fill via the table scan later
-        table_name = self.convert_string_to_table_name(Path(file_uri).stem)
+        table_name = self.convert_string_to_table_name(file_path.stem)
         self.storage.put_object(
             'tables', 
             table_name,
-            {'file_uri': file_uri, 'reader_name': reader.__name__}
+            {'file_uri': file_path.as_uri(), 'reader_name': reader.__name__, 'options': options}
         )
         self.tables = None # force re-calc
         return table_name
 
-    def determine_pandas_reader(self, file_uri: str):
-        res = mimetypes.guess_type(file_uri)
+    def determine_pandas_reader(self, file_uri: Path):
+        res = mimetypes.guess_type(file_uri.as_posix())
         mime = None
         if res[0] is None:
             # Fall back to using 'file' system command
